Rename euclid to coxeter everywhere in the main package and in all top-level files except README.
