+++ .gitignore
@@ -17,3 +17,4 @@ lib
 lib64
 __pycache__
 .hypothesis
+*.pyc
