+++ CONTRIBUTING.md
@@ -3,10 +3,10 @@
 ## Feedback
 
 Issue reports and feature proposals are very welcome.
-Please use the [GitHub issue page](https://github.com/glotzerlab/euclid/issues/) for this.
+Please use the [GitHub issue page](https://github.com/glotzerlab/coexeter/issues/) for this.
 
 ## Contributing code
 
-Code contributions to the euclid open-source project are welcomed via pull requests on [GitHub](https://github.com/glotzerlab/euclid/).
+Code contributions to the coexeter open-source project are welcomed via pull requests on [GitHub](https://github.com/glotzerlab/coexeter/).
 Prior any work you should contact the developers to ensure that the planned development meshes well with the directions and standards of the project.
 All contributors must agree to the Contributor Agreement ([ContributorAgreement.md](ContributorAgreement.md)) before their pull request can be merged.
