+++ LICENSE
@@ -1,6 +1,6 @@
-BSD 3-Clause License for euclid
+BSD 3-Clause License for coexeter
 
-Copyright (c) 2015-2019 The Regents of the University of Michigan All rights reserved.
+Copyright (c) 2015-2020 The Regents of the University of Michigan All rights reserved.
 
 Redistribution and use in source and binary forms, with or without
 modification, are permitted provided that the following conditions are met:
