+++ coxeter/__init__.py
