+++ coxeter/bentley_ottman/LICENSE
