+++ coxeter/bentley_ottman/poly_point_isect.py
