+++ coxeter/common_shapes/RhombicDodecahedron.py
@@ -1,6 +1,6 @@
 from numpy import sqrt
 import numpy
-from euclid.shape_classes.convex_polyhedron import ConvexPolyhedron
+from coxeter.shape_classes.convex_polyhedron import ConvexPolyhedron
 
 points = [
     (-sqrt(2 / 3), -sqrt(2 / 3), 0),
