+++ coxeter/common_shapes/Tetrahedron.py
@@ -1,5 +1,5 @@
 import numpy
-from euclid.shape_classes.convex_polyhedron import ConvexPolyhedron
+from coxeter.shape_classes.convex_polyhedron import ConvexPolyhedron
 
 
 points = [(1, 1, 1), (-1, -1, 1), (1, -1, -1), (-1, 1, -1)]
