+++ coxeter/common_shapes/TruncatedTetrahedron90.py
@@ -1,5 +1,5 @@
 import numpy
-from euclid.shape_classes.convex_polyhedron import ConvexPolyhedron
+from coxeter.shape_classes.convex_polyhedron import ConvexPolyhedron
 
 points = [
     (1, 0.1, 0.1),
