+++ coxeter/common_shapes/__init__.py
