+++ coxeter/damasceno.py
