+++ coxeter/polytri/LICENSE
