+++ coxeter/polytri/polytri.py
