+++ coxeter/shape_classes/__init__.py
