+++ coxeter/shape_classes/circle.py
