+++ coxeter/shape_classes/convex_polygon.py
