+++ coxeter/shape_classes/convex_polyhedron.py
