+++ coxeter/shape_classes/ellipse.py
