+++ coxeter/shape_classes/sphere.py
