+++ coxeter/shapes.py
