+++ coxeter/symmetry.py
