+++ setup.cfg
@@ -4,13 +4,13 @@ commit = True
 tag = True
 message = Bump up to version {new_version}.
 
-[bumpversion:file:euclid/__init__.py]
+[bumpversion:file:coexeter/__init__.py]
 
 [bumpversion:file:doc/source/conf.py]
 
 [bumpversion:file:setup.py]
 
 [flake8]
-exclude = euclid/polytri/*,euclid/bentley_ottman/*
+exclude = coexeter/polytri/*,coexeter/bentley_ottman/*
 per-file-ignores = 
-	euclid/__init__.py: F401
+	coexeter/__init__.py: F401
