+++ setup.py
@@ -20,14 +20,14 @@ extras = {
 }
 
 setup(
-    name='euclid',
+    name='coexeter',
     version='0.1.0',
     description='Tools for creating and manipulating shapes.',
-    url='https://github.com/glotzerlab/euclid',
+    url='https://github.com/glotzerlab/coexeter',
     author='Bryan VanSaders',
     author_email='bvansade@umich.edu',
     packages=find_packages(),
-    package_data={'euclid': ['science.1220869.json']},
+    package_data={'coexeter': ['science.1220869.json']},
     install_requires=[
         'numpy',
         'rowan>=1.2',
